requirements.txt
google-genai==1.5.0
chromadb==0.5.3
PyMuPDF==1.24.5
rank-bm25==0.2.2
python-dotenv==1.0.1
nltk==3.8.1
rouge-score==0.1.2
numpy==1.26.4
pandas==2.2.2
tabulate==0.9.0
tqdm==4.66.4
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
groq